Add files via upload
This is the R code I used to clean the historical data and apply the data with ANSI/ASQ Z1.9 variable sampling method. The method is for standard deviation unknown, and calculated mean, standard deviation, F-value, MSD, M value, QL,QU,PL,PU and P value. And then comparing M value and P to decide if the lot is accepted or not.